computer-maintenance-agent/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psutil==5.9.5
requests==2.31.0
Werkzeug==2.3.7
google-genai
python-dotenv
